# cpms/openapi.yaml
openapi: 3.0.3
info:
  title: CPMS Host Program API
  version: 1.0.0
  description: |
    CPMS 客户端与主机端程序（内网数据服务进程）通信接口。
    仅局域网使用，不暴露互联网。
servers:
  - url: http://cpms-host.local/api/v1
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: 管理员登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
  /auth/refresh:
    post:
      tags: [Auth]
      summary: 刷新访问令牌
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardResponse"
...
  /archives:
    post:
      tags: [Archives]
      summary: 新建档案
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArchiveRequest"
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardResponse"
    get:
      tags: [Archives]
      summary: 档案列表检索
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: archive_index_no
          schema:
            type: string
        - in: query
          name: passport_no
          schema:
            type: string
        - in: query
          name: full_name
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardResponse"
  /archives/{archive_id}:
    get:
      tags: [Archives]
      summary: 档案详情
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: archive_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardResponse"
    put:
      tags: [Archives]
      summary: 更新档案
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: archive_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateArchiveRequest"
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    StandardResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: ok
        data:
          nullable: true
        trace_id:
          type: string
          nullable: true
    LoginResponse:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: ok
        data:
          type: object
          required: [access_token, refresh_token, expires_in, user]
          properties:
            access_token:
              type: string
            refresh_token:
              type: string
            expires_in:
              type: integer
              example: 1800
            user:
              type: object
              required: [user_id, role]
              properties:
                user_id:
                  type: string
                role:
                  type: string
                  enum: [SUPER_ADMIN, ADMIN]
    CreateArchiveRequest:
      type: object
      required: [province_code, full_name, birth_date, gender_code, passport_no]
      properties:
        province_code:
          type: string
          pattern: "^[A-Z]{2}$"
          example: ZS
        full_name:
          type: string
          maxLength: 128
        birth_date:
          type: string
          format: date
          example: "1990-01-15"
        gender_code:
          type: string
          enum: [M, W]
        height_cm:
          type: number
          format: float
          example: 175.5
        passport_no:
          type: string
          maxLength: 32
    UpdateArchiveRequest:
      type: object
      properties:
        full_name:
          type: string
          maxLength: 128
        birth_date:
          type: string
          format: date
        gender_code:
          type: string
          enum: [M, W]
        height_cm:
          type: number
          format: float
        passport_no:
          type: string
          maxLength: 32
        remark:
          type: string